simplified example by removing xdp and batch hash map delete
